Major System Clean-Up
A big code and structure clean-up which should improve overall
performance and to get a better overview
