--- ExempelProjekt/.settings/com.google.gdt.eclipse.core.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-warSrcDir=war
-warSrcDirIsOutput=true
