--- ExempelProjekt/.settings/com.google.gwt.eclipse.core.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-filesCopiedToWebInfLib=gwt-servlet.jar
