--- ExempelProjekt/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<faceted-project>
-  <installed facet="java" version="1.8"/>
-</faceted-project>
