--- ExempelProjekt/src/com/exempel/martin/client/GreetingService.java
@@ -1,12 +1,0 @@
-package com.exempel.martin.client;
-
-import com.google.gwt.user.client.rpc.RemoteService;
-import com.google.gwt.user.client.rpc.RemoteServiceRelativePath;
-
-/**
- * The client-side stub for the RPC service.
- */
-@RemoteServiceRelativePath("greet")
-public interface GreetingService extends RemoteService {
-	String greetServer(String name) throws IllegalArgumentException;
-}
