--- ExempelProjekt/src/com/exempel/martin/client/GreetingServiceAsync.java
@@ -1,11 +1,0 @@
-package com.exempel.martin.client;
-
-import com.google.gwt.user.client.rpc.AsyncCallback;
-
-/**
- * The async counterpart of <code>GreetingService</code>.
- */
-public interface GreetingServiceAsync {
-	void greetServer(String input, AsyncCallback<String> callback)
-			throws IllegalArgumentException;
-}
