+++ GwtCalculator/.settings/com.google.gdt.eclipse.core.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+warSrcDir=war
+warSrcDirIsOutput=true
