+++ GwtCalculator/.settings/com.google.gwt.eclipse.core.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+filesCopiedToWebInfLib=gwt-servlet.jar
