+++ GwtCalculator/.settings/com.gwtplugins.gwt.eclipse.core.prefs
