+++ GwtCalculator/war/WEB-INF/.gitignore
