+++ GwtCalculator/war/calculatorController/gwt/clean/clean_rtl.css
