+++ GwtCalculator/war/calculatorController/gwt/clean/images/circles.png
