+++ GwtCalculator/war/calculatorController/gwt/clean/images/corner.png
