+++ GwtCalculator/war/calculatorController/gwt/clean/images/thumb_vertical.png
