+++ GwtCalculator/war/calculatorController/gwt/clean/images/vborder.png
