+++ GwtCalculator/war/index.html
@@ -1,0 +1,31 @@
+<!doctype html>
+
+<html>
+  <head>
+    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
+   
+    <link type="text/css" rel="stylesheet" href="style.css">
+
+    <title>Final assignment | Webdevelopment 2017</title>
+ 
+    <script type="text/javascript" language="javascript" src="gwtCalculator/gwtCalculator.nocache.js"></script>
+  </head>
+
+  <body>
+      <h1>Autumn 2017 Project - GWT Calculator</h1>
+      <div id="calculator" class="calc"></div>
+      
+      <div id="resultsTable"></div>
+      
+    <!-- RECOMMENDED if your web app will not function without JavaScript enabled -->
+    <noscript>
+      <div style="width: 22em; position: absolute; left: 50%; margin-left: -11em; color: red; background-color: white; border: 1px solid red; padding: 4px; font-family: sans-serif">
+        Your web browser must have JavaScript enabled
+        in order for this application to display correctly.
+      </div>
+    </noscript>
+
+    
+
+      </body>
+</html>
